#87 Close to working backend/frontend
If the frontend with the fetch request is commented out, and we simply use the one that assumes the user is already in the room, it works.

If we use the fetch request, and we have our users sitting in the /chatroom when we start the dev, they are added to the room.  If they try to connect again, they can't get back in.